# expression: mod check null if zero