@hawkingrei 2021-07-10
/run-check_release_note